requirements.txt
os
tkinter
heapq
kivy>=2.1.0
pygame>=2.1.2
Pillow>=9.4.0
numpy>=1.22.4
